refactor > 코드 가독성 및 App컴포넌트 분리
